Fixed scoreboard stopping when a player left and joined
There was a bug that when a player left the game and quickly joined the scoreboard for everyone in the game would stop and not continue the count down. This was due to the player that joined not having a scoreboard so now it will check if they have one and if they don't give them one.